fix: remove note.

+++ src/main/java/org/tron/common/utils/FileUtil.java
@@ -112,7 +112,7 @@ public class FileUtil {
    * delete directory.
    */
   public static boolean deleteDir(File dir) {
-   /* if (dir.isDirectory()) {
+    if (dir.isDirectory()) {
       String[] children = dir.list();
       for (int i = 0; i < children.length; i++) {
         boolean success = deleteDir(new File(dir, children[i]));
@@ -121,7 +121,6 @@ public class FileUtil {
         }
       }
     }
-    return dir.delete();*/
-    return true;
+    return dir.delete();
   }
 }
